move time-related functions out into own module
